References

Contents

References

Contents

Index

API

Index

VTKLegacy.StructuredPoints

VTKLegacy.UnstructuredGrid

Base.getindex

Base.show

VTKLegacy.LoadVTK

VTKLegacy.integrate

VTKLegacy.magnitude

VTKLegacy.magnitude

VTKLegacy.magnitude

VTKLegacy.probe

VTKLegacy.ranges

API

VTKLegacy.StructuredPoints — Type

StructuredPoints

Object that contains all the information and datasets from a Legacy VTK file with a STRUCTURED_POINTS geometry/topology.

Fields

data::Array{Float64,4}: 4-dimensional array of Float64 that holds the datasets of the vtk file.

title::String: Title of the VTK file.

nx::Int32: Number of points in the x direction.

ny::Int32: Number of points in the y direction.

nz::Int32: Number of points in the z direction.

dx::Float64: Distance between points in the x direction.

dy::Float64: Distance between points in the y direction.

dz::Float64: Distance between points in the z direction.

x0::Float64: Position of origin in the x direction.

y0::Float64: Position of origin in the y direction.

z0::Float64: Position of origin in the z direction.

x::Vector{Float64}: Points positions in the x direction.

y::Vector{Float64}: Points positions in the y direction.

z::Vector{Float64}: Points positions in the z direction.

dimensions::Vector{Int64}: Number of points in each dimension: [nx,ny,nz].

spacing::Vector{Float64}: Spacing size in each dimension: [dx,dy,dz].

origin::Vector{Float64}: Origin of the mesh: [x0,y0,z0].

dataNames::Vector{String}: Names of the datasets in the VTK file.

dataAttributes::Vector{String}: Attribute of each dataset in the VTK file.

dictionary::Dict{String,Union{Int64,UnitRange{Int64}}}: Dictionary with the dataNames as the keys and the indexes of data as the values.

VTKLegacy.UnstructuredGrid — Type

UnstructuredGrid

Object that contains all the information and datasets from a Legacy VTK file with a UNSTRUCTURED_GRID geometry/topology.

Fields

title::String: Title of the VTK file.

npoints::Int32: Number of points that define the cells.

ncells::Int32: Number of cells in the file.

points::Array{AbstractFloat,2}: Array containing the coordinates of the points in the grid.

cells::Vector{Int32}: Number and indexes of points defining each cell.

cellTypes::Vector{Int32}: Type of each cell. Integer value between 1 and 16

cellData::Array{AbstractFloat,2}: Array containing each dataset in the CELL_DATA section.

cellDataNames::Vector{String}: Name of each dataset in the CELL_DATA section.

cellDataAttributes::Vector{String}: Attribute of each dataset in the CELL_DATA section.

cellDict::Dict{String,IntOrRng}: Dictionary with the cellDataNames as the keys and the indexes of cellData as the values.

pointData::Array{AbstractFloat,2}: Array containing each dataset in the POINT_DATA section.

pointDataNames::Vector{String}: Name of each dataset in the POINT_DATA section.

pointDataAttributes::Vector{String}: Attribute of each dataset in the POINT_DATA section.

pointDict::Dict{String,IntOrRng}: Dictionary with the pointDataNames as the keys and the indexes of pointData as the values.

Base.getindex — Method

getindex(m::StructuredPoints, name::String)

Retrieve the array stored in the m.data field with name name. The name should be an element of m.dataNames. The syntax m["dataname"] is converted by the compiler to getindex(m,"dataname").

Base.show — Method

show(m::Union{StructuredPoints,UnstructuredGrid})

Print general information of the VTK file contained in the object m

VTKLegacy.LoadVTK — Method

LoadVTK(filename::String)

Read the VTK file filename in to an object of type StructuredPoints or UnstructuredGrid. Currently only StructuredPoints and UnstructuredGrid are supported.

VTKLegacy.integrate — Function

integrate(m::StructuredPoints,var::Union{Int64,String,Vector{Union{Int64,String}}}=1)

Compute the discrete integral of var in the whole mesh m as:

\[I = \sum_{(i,j,k)=(1,1,1)}^{N_x,N_y,N_z}\text{var}_{(i,j,k)}\Delta x\Delta y\Delta z.\]

If var is unspecified, integrate only the dataset with index 1 in m.data

VTKLegacy.magnitude — Method

magnitude(arr::Array{AbstractFloat,3})

Returns the magnitude of a vector dataset of size (axis,nx,ny) where axis is 3 ("x", "y" and "z" dimensions) and nx and ny are the number of cells in each axis.

VTKLegacy.magnitude — Method

magnitude(arr::Array{AbstractFloat,4})

Returns the magnitude of a vector dataset of size (axis,nx,ny,nz) where axis is 3 ("x", "y" and "z" dimensions) and nx, ny and nz are the number of cells in each axis.

VTKLegacy.magnitude — Method

magnitude(m::StructuredPoints,dataname::String)

Returns the magnitude of a dataset in m with name dataname and attribute Vector.

VTKLegacy.probe — Function

probe(filename::String, save::Bool = false, fig = Figure(); kwarg...)

Generate heatmaps of each dataset in the VTK file filename at nz/2. If there are Vector datasets, plot the magnitude of the Vector array. If save is unspecified, the figure will show at the default output stream stdout. Leaving fig unspecified will generate a plot with size (1200,800).

Keyword arguments

image_name::String: When save = true the function will save the plot generated in the same folder filename is at.

axis_slice::String: Accepted values are "x", "y" and "z". Default value is "z".

cell_slice::Int64: Cell at which the slice is created. If nothing is given, the slice will be generated at the midle of the array in the axis given in axis_slice.

VTKLegacy.ranges — Function

ranges(m::StructuredPoints,io::Union{IOStream,Nothing}=nothing)

Obtain the maximum and minimum values of each dataset in m and print them to io followed by a newline.

If io is unspecified, prints the values to the default output stream stdout.